+++ refund.go
@@ -20,7 +20,7 @@ type Refund struct {
 
 type ListRefundResult struct {
 	Refunds []*Refund `json:"refunds,omitempty"`
-	HasMore bool      `json:"has_more,omitempty"`
+	PaginatedCollection
 }
 
 // CreateRefund submits a refund request and return a list of refunds.
@@ -46,6 +46,37 @@ func (c *Client) ListRefunds(opts *ListOptions) (out *ListRefundResult, err erro
 func (c *Client) ListRefundsWithContext(ctx context.Context, opts *ListOptions) (out *ListRefundResult, err error) {
 	err = c.do(ctx, http.MethodGet, "refunds", c.convertOptsToURLValues(opts), &out)
 	return
+}
+
+// GetNextRefundPage returns the next page of refunds
+func (c *Client) GetNextRefundPage(collection *ListRefundResult) (out *ListRefundResult, err error) {
+	return c.GetNextRefundPageWithContext(context.Background(), collection)
+}
+
+// GetNextRefundPageWithPageSize returns the next page of refunds with a specific page size
+func (c *Client) GetNextRefundPageWithPageSize(collection *ListRefundResult, pageSize int) (out *ListRefundResult, err error) {
+	return c.GetNextRefundPageWithPageSizeWithContext(context.Background(), collection, pageSize)
+}
+
+// GetNextRefundPageWithContext performs the same operation as GetNextRefundPage, but
+// allows specifying a context that can interrupt the request.
+func (c *Client) GetNextRefundPageWithContext(ctx context.Context, collection *ListRefundResult) (out *ListRefundResult, err error) {
+	return c.GetNextRefundPageWithPageSizeWithContext(ctx, collection, 0)
+}
+
+// GetNextRefundPageWithPageSizeWithContext performs the same operation as GetNextRefundPageWithPageSize, but
+// allows specifying a context that can interrupt the request.
+func (c *Client) GetNextRefundPageWithPageSizeWithContext(ctx context.Context, collection *ListRefundResult, pageSize int) (out *ListRefundResult, err error) {
+	if len(collection.Refunds) == 0 {
+		err = EndOfPaginationError
+		return
+	}
+	lastID := collection.Refunds[len(collection.Refunds)-1].ID
+	params, err := nextPageParameters(collection.HasMore, lastID, pageSize)
+	if err != nil {
+		return
+	}
+	return c.ListRefundsWithContext(ctx, params)
 }
 
 // retrieves a previously-created Refund by its ID.